requirements.txt
flask
flask_sqlalchemy
mysqlclient
pyserial
MySQL-python
requests